Updated symbolic_link_cleanup (for better or for worse).  Previously, if cleanup involved a symbolic link in the current folder pointing to a file in another folder, then it would attempt to copy (cp) the real file to the a new file with the symbolic link's name.  However the shell would just follow the symbolic link back to the file that the system was trying to copy, resulting in an error instead of the intended action.
The fix is to add a remove (rm) command of the symbolic link before calling the cp command. It has been generalized
to do this as well for the intra-folder link referencing which calls the move (mv) command instead, although it is
not clear that we want to remove it in this case as well.

This trans-folder behaviour is known to be in the set_reference_space_vbm.pm; with the fix we are now going to have
an arguably unnecessary reduncancy of files.  We should consider removing the call to symbolic_link_cleanup from
said module, and allow the space-saving symbolic links to remain.

Also updated in symbolic_link_cleanup is how the logging of the work is handled; the commented code indicates that it
was previously sending the log info to a log that in practice was never fed to the subroutine.  This has been changed
so that it will require the $PM instead (the pipeline code elsewhere was updated accordingly), and will collate all
work messages (including an echo from a command, if any), and use log_info to record them--if there is anything worth
mentioning, that is.